pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PLC"
version = "1.0.1"
authors = [
    { name = "Egor Perevoshchikov", email = "abkein@abkein.ru" },
    { name = "Daniel Mohr", email = "mohr@mpe.mpg.de, daniel.mohr@dlr.de" },
]
description = "PlasmaLabControl is a software for controlling the experiment."
readme = "README.md"
license = { text = "The GNU General Public License v3.0" }
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",

    "License :: OSI Approved :: GNU General Public License (GPL)",

    "Environment :: Console",
    "Environment :: X11 Applications",

    "Programming Language :: Python",
    "Programming Language :: Python :: 3",

    "Operating System :: Unix",
    "Operating System :: POSIX",
    "Operating System :: POSIX :: BSD :: FreeBSD",
    "Operating System :: POSIX :: BSD :: OpenBSD",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",

    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",

    "Natural Language :: English",

    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: System :: Hardware :: Hardware Drivers",
]
dependencies = [
    'ConfigParser',
    'cPickle',
    'cStringIO',
    'distutils.core',
    'errno',
    'importlib',
    'logging',
    'logging.handlers',
    'math',
    'matplotlib',
    'matplotlib.pyplot',
    'numpy',
    'os.path',
    'PIL.ImageOps',
    'PIL.Image',
    'PIL.ImageDraw',
    'PIL.ImageEnhance',
    'PIL.ImageMath',
    'PIL.ImageTk',
    'pydc1394',
    'Queue',
    'serial',
    'SimpleHTTPServer',
    'SocketServer',
    'StringIO',
    'sys',
    'time',
    'tkFileDialog',
    'Tkinter',
    'tkMessageBox',
    'ttk',
    'unittest',
    'usb',
]

# [project.optional-dependencies]
# gui = ["PyQt5"]
# cli = ["rich", "click"]

[project.scripts]

[project.gui-scripts]
plc = "plc"

[project.urls]
"Homepage" = "https://github.com/Architect0rr/MDDPN"
"Bug Tracker" = "https://github.com/Architect0rr/MDDPN/issues"
